pyproject.toml
[project]
name = "agentic-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=0.1.0",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
]
